Feat: [ORC-1439] Create Tests for TNC Workflows

+++ tests/integration_tests/conftest.py
@@ -1,9 +1,15 @@
+# pylint: disable=missing-timeout
 from collections.abc import Callable
 
 import pytest
 import requests
+import structlog
 
 from tests.integration_tests.helpers import API_URL
+
+logger = structlog.get_logger(__name__)
+
+TEST_CLEANUP = True
 
 
 @pytest.fixture
@@ -11,10 +17,12 @@ def products() -> dict[str, str]:
     """Return mapping from product name to product ids"""
     products_response = requests.get(f"{API_URL}/products")
     assert products_response.ok
-    return {product["name"]: product["product_id"] for product in (products_response.json())}
+    return {
+        product["name"]: product["product_id"] for product in (products_response.json())
+    }
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def get_subscriptions() -> Callable:
     def get() -> dict[str, dict]:
         subscriptions_response = requests.get(f"{API_URL}/subscriptions/all")
@@ -24,7 +32,7 @@ def get_subscriptions() -> Callable:
     return get
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def get_subscription_processes():
     def get(subscription_id: str) -> dict[str, dict]:
         processes_response = requests.get(
@@ -36,14 +44,17 @@ def get_subscription_processes():
     return get
 
 
-@pytest.fixture(autouse=True)
+@pytest.fixture(scope="session", autouse=True)
 def clean(get_subscriptions, get_subscription_processes):
     yield
 
-    for sub_id, details in get_subscriptions().items():
-        processes = get_subscription_processes(sub_id)
-        for pid, _ in processes.items():
-            assert requests.delete(f"{API_URL}/processes/{pid}").ok
+    if TEST_CLEANUP:
+        for sub_id, details in get_subscriptions().items():
+            processes = get_subscription_processes(sub_id)
+            for pid, _ in processes.items():
+                assert requests.delete(f"{API_URL}/processes/{pid}").ok
 
-        response = requests.delete(f"{API_URL}/subscriptions/{sub_id}")
-        assert response.ok, f"Failed to delete subscription {sub_id} ({details['description']}): {response.text}"
+            response = requests.delete(f"{API_URL}/subscriptions/{sub_id}")
+            assert (
+                response.ok
+            ), f"Failed to delete subscription {sub_id} ({details['description']}): {response.text}"
